Add WAIT-command (-z)
The WAIT command will periodically check the serial port to detect
magboot presence. If found, the command will succeed and continue
executing remainder of command line.

This is useful if your build environment automatically attempts to
upload new firmware before the device has been reset and magboot is
running.

Under the hood it uses the Device Id command to poll for presence.
If there's no reply in 1s, or if the reply is anything other than
Y, it will try again. This command never fails, it either succeeds
or you cancel it.